Refactor OptiGlue 1.19
