+++ OptiGlue/1.19/src/main/kotlin/opekope2/optiglue_1_19/OptiGlueMod.kt
@@ -1,11 +1,9 @@
-package opekope2.optiglue
+package opekope2.optiglue_1_19
 
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper
 import net.minecraft.entity.vehicle.ChestBoatEntity
 import net.minecraft.resource.ResourceType
 import net.minecraft.util.Nameable
-import opekope2.optiglue.mc_1_19.GlueResource
-import opekope2.optiglue.mc_1_19.RegistryLookupServiceImpl
 import opekope2.optigui.EntryPoint
 import opekope2.optigui.InitializerContext
 import opekope2.optigui.properties.ChestBoatProperties
